backend/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Django==1.11.17
django-apptemplates==1.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
future==0.18.2
idna==2.9
msgpack==1.0.0
mysqlclient==1.4.6
pkg-resources==0.0.0
pop==11
pop-config==5.2
pycparser==2.20
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
serializers==0.2.4
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
